+++ model/vbios.cc
@@ -87,7 +87,7 @@ public:
 
     // initialize realmode idt
     unsigned value = (BIOS_BASE >> 4) << 16;
-    for (unsigned i=0; i < 256; i++) {
+    for (size_t i=0; i < 256; i++) {
       // XXX init only whats needed and done on compatible BIOSes
       if (i != 0x43)
 	discovery_write_dw("realmode idt", i*4, value, 4);
